code/requirements.txt
google-adk==1.1.0
opentelemetry-exporter-gcp-logging==1.9.0a0
opentelemetry-exporter-gcp-monitoring==1.9.0a0
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation-google-genai==0.2b0
opentelemetry-instrumentation-httpx==0.54b1
opentelemetry-instrumentation-requests==0.54b1
opentelemetry-instrumentation-sqlite3==0.54b1
rich==14.0.0
annotated-types==0.7.0
anyio==4.9.0
authlib==1.6.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6 ; sys_platform == 'win32'
cryptography==45.0.3
deprecated==1.2.18
docstring-parser==0.16
fastapi==0.115.12
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.94.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.33.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-monitoring==2.27.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.23.3
google-cloud-speech==2.32.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.1
google-crc32c==1.7.1
google-genai==1.16.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.2 ; (python_full_version < '3.14' and platform_machine == 'AMD64') or (python_full_version < '3.14' and platform_machine == 'WIN32') or (python_full_version < '3.14' and platform_machine == 'aarch64') or (python_full_version < '3.14' and platform_machine == 'amd64') or (python_full_version < '3.14' and platform_machine == 'ppc64le') or (python_full_version < '3.14' and platform_machine == 'win32') or (python_full_version < '3.14' and platform_machine == 'x86_64')
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.6.1
markdown-it-py==3.0.0
mcp==1.9.1
mdurl==0.1.2
numpy==2.2.6
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-dbapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
requests==2.32.3
rsa==4.9.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
sse-starlette==2.3.5
starlette==0.46.2
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2 ; sys_platform == 'win32'
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
psycopg2-binary
pgvector
